Fix issues related to deletions by message count
Before this commit, only the most recent deletable message would be deleted if autodelete was triggered by message count in a channel, and it would delete that message even if it was pinned to the channel. The order of deletion would be most recent first as well; this commit makes deletions by message count happen from oldest to newest.