Migrate configuration to a new, more efficient format
Previously, the value of the "channels" entry in the configuration was a list of channels, each element being a dictionary having a channel ID, time_threshold and max_messages. This required linear time to determine if a channel was set up for auto-delete or not, or to check what a channel's auto-delete configuration was.

This commit introduces a new format for the value of "channels" entry, which is a dictionary that maps channel IDs to dictionaries containing time_threshold and max_messages. This way:
* checking whether a channel is set up for auto-delete becomes the constant-time operation Config.is_channel_set() rather than linear-time;
* getting a single channel's auto-delete settings by ID becomes the constant-time operation Config.get_channel_config() rather than linear-time;
* getting a list of channels set up for auto-delete becomes the constant-time operation Config.get_channels() rather than the linear-time [channel["id"] for channel in channels];
* adding and removing channels no longer traverse a whole list to check whether a channel is set up for auto-delete; those become constant-time operations rather than linear-time;
* iterating over the list of channels set up for auto-delete stays a linear-time operation.

This also paves the way for a database-backed version of Config:
* Config.is_channel_set() could map to SELECT COUNT(*) FROM channels WHERE channel_id = ?;
* Config.get_channel_config() could map to SELECT * FROM channels WHERE channel_id = ?;
* Config.get_channels() could map to SELECT channel_id FROM channels;
* Config.set_channel() could map to INSERT INTO channels (channel_id, time_threshold, max_messages) VALUES (?1, ?2, ?3) ON DUPLICATE KEY UPDATE time_threshold=?2, max_messages=?3;
* Config.clear_channel() could map to DELETE FROM channels WHERE channel_id = ?.

On startup, Melodelete reads the list of dictionaries format and converts it internally to the new single-dictionary format. This new format is saved back to storage after the first change to configuration occurs.